% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnrm_nl}
\alias{cnrm_nl}
\title{Simulation data (CNRM)}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 14610 rows and 2 columns.}
\source{
\href{https://climexp.knmi.nl/select.cgi?id=312456c83e660703df1bfea9ba4fba50&field=pr_A2_cnrm_cm3_20c3m}{KNMI explorer}
}
\usage{
data(cnrm_nl)
}
\description{
Model cnrm-cm3; scenario 20c3m; variable  pr.
24 h 2.8 degree x 2.8 degree for Holland
at daily time step for period 1961-01-01 to 2000-12-31.
Spatial Region: 1 grid cell at latitude: 51.625, longitude: 5.625
}
\examples{
str(cnrm_nl)
}
\keyword{datasets}
