% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{csa.rescale}
\alias{csa.rescale}
\title{Changes the scale of an aggregation curve}
\usage{
csa.rescale(csa_coarse, csa_fine, scale_ratio)
}
\value{
as a brick object.
}
\description{
The function \code{csa.rescale} needs to be desribed.
}
