% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{dt.to.brick}
\alias{dt.to.brick}
\title{Trasnform data.table to brick}
\usage{
dt.to.brick(dt, var_name)
}
\arguments{
\item{dt}{The data table object to be transformed. It must be in a four-column format, with the coordinate columns named as "lat" & "lon"
and time values as "time".}

\item{var_name}{The name (chr) of the column in the data table (\code{dt}) wich holds the values of the variable, e.g., "temperature".}
}
\value{
\code{dt} as a brick object.
}
\description{
The function \code{dt.to.brick} tranforms a data.table object to brick (raster) format
}
