% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpm_events}
\alias{gpm_events}
\title{GPM-IMERG precipitation events over 10 mm/day}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6612 rows and 6 columns.
}
\source{
\href{https://climexp.knmi.nl/select.cgi?id=312456c83e660703df1bfea9ba4fba50&field=imerg_daily}{KNMI explorer}
}
\usage{
data(gpm_events)
}
\description{
GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree for Holland
at daily time step for period 2014-03-12 to 2018-05-15.
Spatial averaged over: latitude: 50.75, 53.55, longitude: 3.45, 7.15
}
\examples{
str(gpm_events)
}
\keyword{datasets}
