% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpm_nl}
\alias{gpm_nl}
\title{Reconstructed summer drought for Czech Republic}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1526 rows and 2 columns.}
\usage{
gpm_nl
}
\description{
Grided reconstruction (0.5 x 0.5 deg) of the self-calibrated
Palmer Drought Index for summer (JJA) precipitation over Europe, as presented
by Cook et al. [2015] and edited by Markonis et al. [2017]. Here, only the
area for Czech Republic is used (49 - 51 deg N and 13 - 17 E).
}
\examples{
## Plot scalegram in logarithmic y axis
data(gpm_nl)
}
\references{
Cook, E. R., Seager, R., Kushnir, Y., Briffa, K. R., Büntgen, U., Frank, D., ... & Baillie, M. (2015).
Old World megadroughts and pluvials during the Common Era. Science advances, 1(10).

Markonis, Y., Hanel, M., Maca, P., Kysely, Y., & Cook E. (2017). Hydroclimatic conditions of Europe in a millennial perspective.
Nature Communications, X(X).
}
\keyword{datasets}
