% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knmi_nl}
\alias{knmi_nl}
\title{Station data (KNMI)}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 24592 rows and 2 columns.}
\source{
\href{https://climexp.knmi.nl/PhomNL.cgi?id=312456c83e660703df1bfea9ba4fba50}{KNMI explorer}
}
\usage{
data(knmi_nl)
}
\description{
240 homogenized stations 1951-now.
24 h point data for Holland
at daily time step for period 1950-12-31 to 2018-04-29.
Spatial Region: latitude: 50.78, 53.48, longitude: 3.4, 7.11
}
\examples{
str(knmi_nl)
}
\keyword{datasets}
