% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncep_nl}
\alias{ncep_nl}
\title{Reanalysis data (NCEP/NCAR)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 25601 rows and 2 columns.
}
\source{
\href{https://climexp.knmi.nl/select.cgi?id=312456c83e660703df1bfea9ba4fba50&field=nprate_daily}{KNMI explorer}
}
\usage{
data(ncep_nl)
}
\description{
NMC reanalysis 24 h 2.5 degree x 2.5 degree for Holland
at daily time step for period 1948-01-01 to 2018-06-05.
Spatial Region: 1 grid cell at latitude: 52.38, longitude: 5.625
}
\examples{
str(ncep_nl)
}
\keyword{datasets}
