% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdr_nl}
\alias{rdr_nl}
\title{Radar data (KNMI)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1472 rows and 2 columns.
}
\source{
\href{https://climexp.knmi.nl/select.cgi?id=312456c83e660703df1bfea9ba4fba50&field=knmi_radar_daily}{KNMI explorer}
}
\usage{
data(rdr_nl)
}
\description{
RAD_NL25_RAC_MFBS_24H_NC
24 h 1 km x 1 km for Holland
at daily time step for period 2014-03-11 to 2018-03-30.
Spatial Region: latitude: 50.76, 53.56, longitude: 3.37, 7.22
}
\examples{
str(rdr_nl)
}
\keyword{datasets}
